Convert classes in StmtOpenMP to use TrailingObjects.
These had been a bit unique before, in that they used
llvm::RoundUpToAlignment to ensure alignment, rather than ensuring
alignment of the base type directly as was done in other parts of the
code.

These classes are still a bit of a unique snowflake, with the storage of
ClausesOffset in the base class to store the end of the derived
object (vs e.g. OMPVarListClause which causes the accessors to just be
defined properly for each subclass), and the use of the trailing objects
idiom for storing non-varying numbers of objects.