Default vaarg lowering should support indirect struct types.
On most targets, clang lowers va_arg itself, eschewing the use of the
llvm vaarg instruction. This is necessary, as the argument to the vaarg
instruction does not represent all the ABI information that is needed to
support complex calling conventions.

However, on targets with a simpler varrags ABIs, the LLVM instruction
can work just fine, and clang can simply lower to it. Unfortunately,
this meant that on such targets, vaarg with a struct argument would
fail, because the default lowering to vaarg was naive: it didn't take
account of the ABI attribute computed by classifyArgumentType. In
particular, for the DefaultABIInfo, structs are passed indirectly and so
llvm's vaarg instruction should have been emitted with a pointer
argument.

Now, vaarg instruction emission is able to use computed ABIArgInfo for
the provided argument type, which allows the default ABI support to work
for structs too.

I've not removed the EmitVAArg implementation for PPC32_SVR4 or XCore,
although I believe both are now redundant, and could be switched over to
use the default implementation as well.